minor fixes and adds to shape package
